Hotfix

+++ src/main/java/org/springframework/samples/petclinic/web/PetController.java
@@ -15,27 +15,31 @@
  */
 package org.springframework.samples.petclinic.web;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.samples.petclinic.model.Owner;
-import org.springframework.samples.petclinic.model.Pet;
-import org.springframework.samples.petclinic.model.PetType;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.ModelMap;
-import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.WebDataBinder;
-import org.springframework.web.bind.annotation.*;
-
-import javax.validation.Valid;
-
 import java.util.Collection;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+import javax.validation.Valid;
+
 import org.springframework.beans.BeanUtils;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.samples.petclinic.model.Owner;
+import org.springframework.samples.petclinic.model.Pet;
+import org.springframework.samples.petclinic.model.PetType;
 import org.springframework.samples.petclinic.service.OwnerService;
 import org.springframework.samples.petclinic.service.PetService;
 import org.springframework.samples.petclinic.service.exceptions.DuplicatedPetNameException;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.ModelMap;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.WebDataBinder;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.InitBinder;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 /**
  * @author Juergen Hoeller
@@ -45,16 +49,18 @@ import org.springframework.samples.petclinic.service.exceptions.DuplicatedPetNam
 @Controller
 @RequestMapping("/owners/{ownerId}")
 public class PetController {
+	
+	private static final String OWNERS_DETAILS_REDIRECT = "redirect:/owners/{ownerId}";
 
 	private static final String VIEWS_PETS_CREATE_OR_UPDATE_FORM = "pets/createOrUpdatePetForm";
 
 	private final PetService petService;
-    private final OwnerService ownerService;
+	private final OwnerService ownerService;
 
 	@Autowired
 	public PetController(final PetService petService, final OwnerService ownerService) {
 		this.petService = petService;
-        this.ownerService = ownerService;
+		this.ownerService = ownerService;
 	}
 
 	@ModelAttribute("types")
@@ -66,17 +72,7 @@ public class PetController {
 	public Owner findOwner(@PathVariable("ownerId") final int ownerId) {
 		return this.ownerService.findOwnerById(ownerId);
 	}
-        
-        /*@ModelAttribute("pet")
-	public Pet findPet(@PathVariable("petId") Integer petId) {
-            Pet result=null;
-		if(petId!=null)
-                    result=this.clinicService.findPetById(petId);
-                else
-                    result=new Pet();
-            return result;
-	}*/
-                
+
 	@InitBinder("owner")
 	public void initOwnerBinder(final WebDataBinder dataBinder) {
 		dataBinder.setDisallowedFields("id");
@@ -96,21 +92,21 @@ public class PetController {
 	}
 
 	@PostMapping(value = "/pets/new")
-	public String processCreationForm(final Owner owner, @Valid final Pet pet, final BindingResult result, final ModelMap model) {		
+	public String processCreationForm(final Owner owner, @Valid final Pet pet, final BindingResult result,
+			final ModelMap model) {
 		if (result.hasErrors()) {
 			model.put("pet", pet);
 			return VIEWS_PETS_CREATE_OR_UPDATE_FORM;
-		}
-		else {
-                    try{
-                    	owner.addPet(pet);
-                    	pet.setEnAdopcion(false);
-                    	this.petService.savePet(pet);
-                    }catch(final DuplicatedPetNameException ex){
-                        result.rejectValue("name", "duplicate", "already exists");
-                        return VIEWS_PETS_CREATE_OR_UPDATE_FORM;
-                    }
-                    return "redirect:/owners/{ownerId}";
+		} else {
+			try {
+				owner.addPet(pet);
+				pet.setEnAdopcion(false);
+				this.petService.savePet(pet);
+			} catch (final DuplicatedPetNameException ex) {
+				result.rejectValue("name", "duplicate", "already exists");
+				return VIEWS_PETS_CREATE_OR_UPDATE_FORM;
+			}
+			return OWNERS_DETAILS_REDIRECT;
 		}
 	}
 
@@ -120,46 +116,48 @@ public class PetController {
 		model.put("pet", pet);
 		return VIEWS_PETS_CREATE_OR_UPDATE_FORM;
 	}
-	
+
 	@GetMapping(value = "/pets/{petId}/delete")
 	public String deletePet(@PathVariable("petId") final int petId, final ModelMap model) {
 		final Pet pet = petService.findPetById(petId);
 		final Owner owner = (Owner) model.getAttribute("owner");
-		final Set<Pet> pets = new HashSet<Pet>(owner.getPets().stream().collect(Collectors.toSet()));
-		pets.remove(pet);
-		owner.setPets(pets);
-		ownerService.save(owner);
-		petService.deleteById(petId);
-		return "redirect:/owners/{ownerId}";
+		if (owner != null) {
+			final Set<Pet> pets = new HashSet<>(owner.getPets().stream().collect(Collectors.toSet()));
+			pets.remove(pet);
+			owner.setPets(pets);
+			ownerService.save(owner);
+			petService.deleteById(petId);
+		}
+		return OWNERS_DETAILS_REDIRECT;
 	}
 
-    /**
-     *
-     * @param pet
-     * @param result
-     * @param petId
-     * @param model
-     * @param owner
-     * @param model
-     * @return
-     */
-        @PostMapping(value = "/pets/{petId}/edit")
-	public String processUpdateForm(@Valid final Pet pet, final BindingResult result, final Owner owner,@PathVariable("petId") final int petId, final ModelMap model) {
+	/**
+	 *
+	 * @param pet
+	 * @param result
+	 * @param petId
+	 * @param model
+	 * @param owner
+	 * @param model
+	 * @return
+	 */
+	@PostMapping(value = "/pets/{petId}/edit")
+	public String processUpdateForm(@Valid final Pet pet, final BindingResult result, final Owner owner,
+			@PathVariable("petId") final int petId, final ModelMap model) {
 		if (result.hasErrors()) {
 			model.put("pet", pet);
 			return VIEWS_PETS_CREATE_OR_UPDATE_FORM;
+		} else {
+			final Pet petToUpdate = this.petService.findPetById(petId);
+			BeanUtils.copyProperties(pet, petToUpdate, "id", "owner", "visits");
+			try {
+				this.petService.savePet(petToUpdate);
+			} catch (final DuplicatedPetNameException ex) {
+				result.rejectValue("name", "duplicate", "already exists");
+				return VIEWS_PETS_CREATE_OR_UPDATE_FORM;
+			}
+			return OWNERS_DETAILS_REDIRECT;
 		}
-		else {
-                        final Pet petToUpdate=this.petService.findPetById(petId);
-			BeanUtils.copyProperties(pet, petToUpdate, "id","owner","visits");                                                                                  
-                    try {                    
-                        this.petService.savePet(petToUpdate);                    
-                    } catch (final DuplicatedPetNameException ex) {
-                        result.rejectValue("name", "duplicate", "already exists");
-                        return VIEWS_PETS_CREATE_OR_UPDATE_FORM;
-                    }
-			return "redirect:/owners/{ownerId}";
-		}
-	
-     }
+
+	}
 }
